Add Timer component
